requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==3.0.0
channels_redis==4.2.1
click==8.1.7
constantly==23.10.4
cryptography==44.0.0
daphne==3.0.2
Django==3.1.12
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.7.0
djongo==1.3.7
dnspython==2.7.0
drf-spectacular==0.27.2
geographiclib==2.0
geopy==2.4.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
mongoengine==0.29.1
msgpack==1.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.0
pymongo==3.11.4
pyOpenSSL==24.3.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
rpds-py==0.21.0
service-identity==24.2.0
sqlparse==0.2.4
tomli==2.1.0
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
zope.interface==7.2
